+++ exporter/opentelemetry-exporter-otlp-proto-common/tests/test_metrics_encoder.py
@@ -171,7 +171,7 @@ class TestOTLPMetricsEncoder(unittest.TestCase):
                                                 start_time_unix_nano=1641946015139533244,
                                                 time_unix_nano=1641946016139533244,
                                                 as_int=33,
-                                                flags=DataPointFlags.DEFAULT,
+                                                flags=int(DataPointFlags.get_default()),
                                             )
                                         ],
                                         aggregation_temporality=AggregationTemporality.CUMULATIVE,
@@ -524,7 +524,7 @@ class TestOTLPMetricsEncoder(unittest.TestCase):
                                                 ],
                                                 max=18.0,
                                                 min=8.0,
-                                                flags=DataPointFlags.DEFAULT,
+                                                flags=int(DataPointFlags.get_default()),
                                             )
                                         ],
                                         aggregation_temporality=AggregationTemporality.DELTA,
@@ -656,7 +656,7 @@ class TestOTLPMetricsEncoder(unittest.TestCase):
                                                 ],
                                                 max=18.0,
                                                 min=8.0,
-                                                flags=DataPointFlags.DEFAULT,
+                                                flags=int(DataPointFlags.get_default()),
                                             )
                                         ],
                                         aggregation_temporality=AggregationTemporality.DELTA,
@@ -719,7 +719,7 @@ class TestOTLPMetricsEncoder(unittest.TestCase):
                                                 ],
                                                 max=18.0,
                                                 min=8.0,
-                                                flags=DataPointFlags.DEFAULT,
+                                                flags=int(DataPointFlags.get_default()),
                                             )
                                         ],
                                         aggregation_temporality=AggregationTemporality.DELTA,
@@ -790,7 +790,7 @@ class TestOTLPMetricsEncoder(unittest.TestCase):
                                                 ],
                                                 max=18.0,
                                                 min=8.0,
-                                                flags=DataPointFlags.DEFAULT,
+                                                flags=int(DataPointFlags.get_default()),
                                             )
                                         ],
                                         aggregation_temporality=AggregationTemporality.DELTA,
@@ -861,7 +861,7 @@ class TestOTLPMetricsEncoder(unittest.TestCase):
                                                 ],
                                                 max=18.0,
                                                 min=8.0,
-                                                flags=DataPointFlags.DEFAULT,
+                                                flags=int(DataPointFlags.get_default()),
                                             )
                                         ],
                                         aggregation_temporality=AggregationTemporality.DELTA,
@@ -978,7 +978,7 @@ class TestOTLPMetricsEncoder(unittest.TestCase):
                                                     offset=9,
                                                     bucket_counts=[10, 11],
                                                 ),
-                                                flags=DataPointFlags.DEFAULT,
+                                                flags=int(DataPointFlags.get_default()),
                                                 exemplars=[],
                                                 min=13.0,
                                                 max=14.0,